生成发货任务

+++ base-demo/src/main/java/org/example/o_mysql/domain/TIssueLine.java
@@ -37,7 +37,7 @@ public class TIssueLine implements Serializable {
      * 任务数量
      */
     @TableField(value = "get_num")
-    private Integer getNum;
+    private Double getNum;
 
     /**
      * 状态
